+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added the ability to implement `SingleChildWidget` using mixins
+- Added support for `InheritedWidget` (only using mixins).
+
 ## 0.0.1
 
 Initial release